fix(fastify): forward raw request to node-http handler
Two weeks ago, there was a change on `trpc/server` package that added
usage of `req.off()` in `adapters/node-http/incomingMessageToRequest.ts`.

See relatable:

- https://github.com/trpc/trpc/pull/6842

This change put in light that the trpc-to-openapi Fastify adapter was
incorrectly passing the _fastify_ parsed request to node-http-handler,
and this object does not have `req.off()`, which resulted in a runtime
error.

Example stacktrace from our app:

```
<redacted>/node_modules/.pnpm/@trpc+server@11.4.3_typescript@5.8.3/node_modules/@trpc/server/dist/node-http-CCJdFkum.cjs:74
		req.off("aborted", onAbort);
		   ^

TypeError: req.off is not a function
    at ServerResponse.onAbort (<redacted>/node_modules/.pnpm/@trpc+server@11.4.3_typescript@5.8.3/node_modules/@trpc/server/dist/node-http-CCJdFkum.cjs:74:7)
    at Object.onceWrapper (node:events:621:28)
    at ServerResponse.emit (node:events:507:28)
    at emitCloseNT (node:_http_server:1026:10)
    at process.processTicksAndRejections (node:internal/process/task_queues:89:21)

Node.js v23.1.0
```

Hence, this PR is doing two things:

- pass `request.raw` to `openApiHttpHandler`, so the types are correct
- assign `request.body` (the body parsed by Fastify) to `request.raw`,
  so it properly get handled in `getBody()` in
  `src/adapters/node-http/input.ts` (not doing that was resulting in an
  error because `setEncoding()` was already set on the stream).

I could validate that the change is solving the issue on our side, and
does not seem to break any test.

Let me know if this could cause un-wanted side effect.

Funnily, this looks similar to a past PR from me few months ago
:sweat_smile: :

- https://github.com/mcampa/trpc-to-openapi/pull/63

Thanks for maintaining this!